requirements.txt
aiogram==3.20.0
python-dotenv==1.0.*
